Move object matchers to the machers util file, rename worker
